Add client trait
